+++ src/hubmap_ingest_validator/validator.py
@@ -13,12 +13,14 @@ class TableValidationErrors(Exception):
     pass
 
 
-def validate(path, type, donor_id, tissue_id):
+def validate(path, type, donor_id, tissue_id, skip_data_path=False):
     path_obj = Path(path)
     _validate_generic_submission(path_obj)
     _validate_dataset_directories(path_obj, type)
     _validate_metadata_tsv(path_obj / 'metadata.tsv', type.split('-')[0])
     _validate_references_up(path_obj / 'metadata.tsv', donor_id, tissue_id)
+    if not skip_data_path:
+        _validate_references_down(path_obj)
 
 
 def _validate_generic_submission(dir_path):
@@ -79,3 +81,7 @@ def _validate_references_up(metadata_path, donor_id, tissue_id):
                     f'On row {i+1}, tissue_id is "{row["tissue_id"]}"')
     if error_messages:
         raise TableValidationErrors('\n'.join(error_messages))
+
+
+def _validate_references_down(dir_path):
+    pass