+++ src/validate.py
@@ -42,6 +42,9 @@ def main():
         '--tissue_id', type=str, metavar='ID', required=True,
         help='HuBMAP Display ID of Tissue')
     parser.add_argument(
+        '--skip_data_path', action='store_true',
+        help='If present, the data_path will not be validated')
+    parser.add_argument(
         '--logging', metavar='LOG_LEVEL', type=str,
         choices=['DEBUG', 'INFO', 'WARN'],
         default='WARN')
@@ -49,14 +52,17 @@ def main():
     logging.basicConfig(level=args.logging)
     return _print_message(
         args.dir, args.type,
-        args.donor_id, args.tissue_id
+        args.donor_id, args.tissue_id,
+        skip_data_path=args.skip_data_path
     )
 
 
-def _print_message(dir, type, donor_id, tissue_id, periods=False):
+def _print_message(
+        dir, type, donor_id, tissue_id,
+        periods=False, skip_data_path=False):
     # Doctests choke on blank lines: periods=True replaces with "." for now.
     try:
-        validate(dir, type, donor_id, tissue_id)
+        validate(dir, type, donor_id, tissue_id, skip_data_path=skip_data_path)
         logging.info('PASS')
         return 0
     except DirectoryValidationErrors as e: